+++ src/Services/ClientFactory.php
@@ -1,10 +1,12 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Drupal\opening_hours\Services;
 
 use Drupal\Core\Config\ConfigFactoryInterface;
 use GuzzleHttp\Client as GuzzleClient;
-use StadGent\Services\OpeningHours\Client\Client as StadGentClient;
+use StadGent\Services\OpeningHours\Client\Client;
 use StadGent\Services\OpeningHours\Configuration\Configuration;
 
 /**
@@ -12,7 +14,7 @@ use StadGent\Services\OpeningHours\Configuration\Configuration;
  *
  * @package Drupal\opening_hours\Services
  */
-class ClientService extends StadGentClient {
+final class ClientFactory {
 
   /**
    * Class constructor.
@@ -20,7 +22,7 @@ class ClientService extends StadGentClient {
    * @param \Drupal\Core\Config\ConfigFactoryInterface $config_factory
    *   The config factory.
    */
-  public function __construct(ConfigFactoryInterface $config_factory) {
+  public static function create(ConfigFactoryInterface $config_factory): Client {
     $config = $config_factory->get('opening_hours.settings');
 
     $configuration = new Configuration(
@@ -30,7 +32,7 @@ class ClientService extends StadGentClient {
 
     $guzzleClient = new GuzzleClient(['base_uri' => $configuration->getUri()]);
 
-    parent::__construct($guzzleClient, $configuration);
+    return new Client($guzzleClient, $configuration);
   }
 
 }