+++ CHANGELOG.md
@@ -2,6 +2,28 @@
 
 All Notable changes to `drupal/opening-hours` module.
 
+## [2.0.0]
+
+### Fixed
+
+- Fix code issues.
+
+### Changed
+
+- Change minimal drupal version to 9.3.
+- Change Client service creation to a factory.
+
+### Updated
+
+- Update qa-drupal.
+- Update stadgent/services-opening-hours package to 2.x.
+
+## [1.5.0]
+
+### Added
+
+- Add support for PHP 8.x.
+
 ## [1.4.1]
 
 ### Changed
@@ -213,6 +235,8 @@ for the same widget.
 * DMOH-20: Added the opening hours field type.
 * DMOH-21: Added the opening hours field widget.
 
+[2.0.0]: https://github.com/StadGent/drupal_module_opening-hours/compare/1.5.0...2.0.0
+[1.5.0]: https://github.com/StadGent/drupal_module_opening-hours/compare/1.4.1...1.5.0
 [1.4.1]: https://github.com/StadGent/drupal_module_opening-hours/compare/1.4.0...1.4.1
 [1.4.0]: https://github.com/StadGent/drupal_module_opening-hours/compare/8.x-1.3...1.4.0
 [8.x-1.3]: https://github.com/StadGent/drupal_module_opening-hours/compare/8.x-1.2...8.x-1.3
@@ -236,4 +260,4 @@ for the same widget.
 [8.x-1.0-alpha3]: https://github.com/StadGent/drupal_module_opening-hours/compare/8.x-1.0-alpha2...8.x-1.0-alpha3
 [8.x-1.0-alpha2]: https://github.com/StadGent/drupal_module_opening-hours/compare/8.x-1.0-alpha1...8.x-1.0-alpha2
 [8.x-1.0-alpha1]: https://github.com/StadGent/drupal_module_opening-hours/releases/tag/8.x-1.0-alpha1
-[Unreleased]: https://github.com/StadGent/drupal_module_opening-hours/compare/master...develop
+[Unreleased]: https://github.com/StadGent/drupal_module_opening-hours/compare/main...develop